#58 split pmodel classes into separate files
