+++ pyproject.toml
@@ -27,6 +27,7 @@ numpy = "^1.16.5"
 scipy = "^1.7.3"
 tabulate = "^0.8.10"
 dacite = "^1.6.0"
+graphviz = "*"
 
 [tool.poetry.group.types.dependencies]
 types-tabulate = "^0.9.0.0"
